16th Commit: Further safety save, I am refining and further making great improvements to the ECS.

+++ Scripts/Components/Harmable.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 namespace TheRuinsOfIpsus
 {
     [Serializable]
-    public class OnHit: Component
+    public class Harmable: Component
     {
         public List<string> statusEffects = new List<string>();
         public void Hit(int dmg, string type, string weaponName, Entity attacker) 
@@ -29,7 +29,7 @@ namespace TheRuinsOfIpsus
 
             if (stats.hp <= 0) 
             {
-                Death();
+                Death(attacker.GetComponent<Description>().name);
             }
             else
             {
@@ -38,7 +38,7 @@ namespace TheRuinsOfIpsus
                 {
                     Log.Add($"{attacker.GetComponent<Description>().name} hit {pronounSet.reflexive} for {dmg} damage with {pronounSet.possesive} {weaponName}.");
                 }
-                else if (pronounSet != null && !pronounSet.present)
+                else if (pronounSet != null && pronounSet.present)
                 {
                     Log.Add($"{attacker.GetComponent<Description>().name} hit {entity.GetComponent<Description>().name} for {dmg} damage with {pronounSet.possesive} {weaponName}."); 
                 }
@@ -61,7 +61,7 @@ namespace TheRuinsOfIpsus
                 Renderer.AddParticle(vector2.x, vector2.y, hitParticle);
             }
         }
-        public void LowerHealth(int dmg)
+        public void LowerHealth(int dmg, string cause)
         {
             Stats stats = entity.GetComponent<Stats>();
             stats.hp -= dmg;
@@ -71,28 +71,37 @@ namespace TheRuinsOfIpsus
             }
             if (stats.hp <= 0) 
             {
-                Death(); 
+                Death(cause); 
             }
         }
-        public void Death() 
+        public void Death(string causeOfDeath) 
         {
-            Entity hitParticle = new Entity(new List<Component>
+            if (!entity.display)
+            {
+                Entity hitParticle = new Entity(new List<Component>
                         {
                             new Coordinate(0, 0),
                             new Draw("Red", "Black", 'X'),
                             new ParticleComponent(2, 1, "None", 1, new Draw[] { new Draw("Red", "Black", 'X') })
                         });
-            Coordinate vector2 = entity.GetComponent<Coordinate>();
-            Renderer.AddParticle(vector2.vector2.x, vector2.vector2.y, hitParticle);
-            Log.Add($"{entity.GetComponent<Description>().name} has died.");
-            entity.GetComponent<TurnFunction>().turnActive = false;
+                Coordinate vector2 = entity.GetComponent<Coordinate>();
+                Renderer.AddParticle(vector2.vector2.x, vector2.vector2.y, hitParticle);
+                Log.Add($"{entity.GetComponent<Description>().name} has died.");
 
-            World.GetTraversable(entity.GetComponent<Coordinate>().vector2).actorLayer = null; 
-            TurnManager.RemoveActor(entity.GetComponent<TurnFunction>());
-            EntityManager.RemoveEntity(entity);
+                entity.RemoveComponent(CMath.ReturnAI(entity));
+                TurnManager.RemoveActor(entity.GetComponent<TurnFunction>());
+                EntityManager.ClearAIOfEntity(entity);
+                entity.GetComponent<TurnFunction>().turnActive = false;
+                World.GetTraversable(entity.GetComponent<Coordinate>().vector2).actorLayer = null;
+                EntityManager.RemoveEntity(entity);
 
-            entity.ClearEntity();
+                entity.ClearEntity();
+            }
+            else
+            {
+                Menu.EndGame(causeOfDeath);
+            }
         }
-        public OnHit() { }
+        public Harmable() { }
     }
 }
